Add login system and posting functionality
