+++ backend/db/models.py
@@ -1,0 +1,69 @@
+# backend/db/models.py
+from sqlalchemy import Boolean, Column, ForeignKey, Integer, String, DateTime, Text, Enum
+from sqlalchemy.orm import relationship
+from sqlalchemy.sql import func
+import enum
+from uuid import uuid4
+from .database import Base
+
+class ContentType(str, enum.Enum):
+    document = "document"
+    image = "image"
+    video = "video"
+
+class ProjectCategory(str, enum.Enum):
+    comics = "comics"
+    animation = "animation"
+    novel = "novel"
+    script = "script"
+    ai = "ai"
+    other = "other"
+
+class User(Base):
+    __tablename__ = "users"
+
+    id = Column(Integer, primary_key=True, index=True)
+    username = Column(String, unique=True, index=True)
+    email = Column(String, unique=True, index=True)
+    full_name = Column(String)
+    hashed_password = Column(String)
+    bio = Column(Text, nullable=True)
+    is_active = Column(Boolean, default=True)
+    created_at = Column(DateTime(timezone=True), server_default=func.now())
+    updated_at = Column(DateTime(timezone=True), onupdate=func.now())
+
+    # Relationships
+    projects = relationship("Project", back_populates="owner")
+    posts = relationship("Post", back_populates="author")
+
+class Project(Base):
+    __tablename__ = "projects"
+
+    id = Column(Integer, primary_key=True, index=True)
+    title = Column(String, index=True)
+    description = Column(Text)
+    category = Column(String, default="other")
+    owner_id = Column(Integer, ForeignKey("users.id"))
+    created_at = Column(DateTime(timezone=True), server_default=func.now())
+    updated_at = Column(DateTime(timezone=True), onupdate=func.now())
+
+    # Relationships
+    owner = relationship("User", back_populates="projects")
+    posts = relationship("Post", back_populates="project")
+
+class Post(Base):
+    __tablename__ = "posts"
+
+    id = Column(Integer, primary_key=True, index=True)
+    title = Column(String, index=True)
+    content = Column(Text)
+    content_type = Column(String, default="document")
+    media_url = Column(String, nullable=True)
+    project_id = Column(Integer, ForeignKey("projects.id"))
+    author_id = Column(Integer, ForeignKey("users.id"))
+    created_at = Column(DateTime(timezone=True), server_default=func.now())
+    updated_at = Column(DateTime(timezone=True), onupdate=func.now())
+
+    # Relationships
+    project = relationship("Project", back_populates="posts")
+    author = relationship("User", back_populates="posts")
