+++ backend/db/schemas.py
@@ -1,0 +1,65 @@
+# backend/db/schemas.py
+from typing import List, Optional
+from pydantic import BaseModel, EmailStr
+from datetime import datetime
+
+# User schemas
+class UserBase(BaseModel):
+    email: EmailStr
+    username: str
+    full_name: Optional[str] = None
+
+class UserCreate(UserBase):
+    password: str
+
+class UserOut(UserBase):
+    id: int
+    is_active: bool
+    created_at: datetime
+
+    class Config:
+        from_attributes = True
+
+# Token schemas
+class Token(BaseModel):
+    access_token: str
+    token_type: str
+
+class TokenData(BaseModel):
+    user_id: Optional[str] = None
+
+# Project schemas
+class ProjectBase(BaseModel):
+    title: str
+    description: str
+    category: Optional[str] = "other"
+
+class ProjectCreate(ProjectBase):
+    pass
+
+class Project(ProjectBase):
+    id: int
+    owner_id: int
+    created_at: datetime
+
+    class Config:
+        from_attributes = True
+
+# Post schemas
+class PostBase(BaseModel):
+    title: str
+    content: str
+    content_type: str = "document"
+    project_id: int
+
+class PostCreate(PostBase):
+    pass
+
+class Post(PostBase):
+    id: int
+    author_id: int
+    created_at: datetime
+    media_url: Optional[str] = None
+
+    class Config:
+        from_attributes = True
