+++ backend/requirements.txt
@@ -1,14 +1,15 @@
 # backend/requirements.txt
-fastapi==0.103.1
-uvicorn==0.23.2
-sqlalchemy==2.0.20
-pydantic==2.3.0
-pydantic-extra-types==2.1.0
-pydantic-settings==2.0.3
+fastapi==0.110.0
+uvicorn==0.29.0
+sqlalchemy==2.0.29
+pydantic==2.6.4
+pydantic-extra-types==2.6.0
+pydantic-settings==2.1.0
 python-jose==3.3.0
 passlib==1.7.4
-bcrypt==4.0.1
-python-multipart==0.0.6
-email-validator==2.0.0.post2
-psycopg2-binary==2.9.7
-python-dotenv==1.0.0
+bcrypt==4.1.2
+python-multipart==0.0.9
+email-validator==2.1.0.post1
+psycopg2-binary==2.9.9
+python-dotenv==1.0.1
+PyJWT==2.8.0
